move parts of the returns into separate components
